% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{add100}
\alias{add100}
\title{Add 100 to the given number}
\usage{
add100(x)
}
\arguments{
\item{x}{a number, float or int}
}
\value{
a number
}
\description{
Add 100 to the given number
}
\examples{
add100(5)
add100(-90)
}
