% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{my_pie}
\alias{my_pie}
\title{Using a specific column of a data frame to plot a pie plot}
\usage{
my_pie(data, ncol, title = "")
}
\arguments{
\item{data}{a dataframe or tibble}

\item{ncol}{specific column, factor}
}
\value{
a figure as ggplot2 object
}
\description{
Using a specific column of a data frame to plot a pie plot
}
\examples{
my_pie(mtcars,ncol=2)
my_pie(diamonds,ncol=3)
my_pie(diamonds,ncol=3)+guides(fill="none")
}
