% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{my_pie2}
\alias{my_pie2}
\title{Using a specific column of a data frame to plot a pie plot-v2}
\usage{
my_pie2(data, ncol, title = "")
}
\arguments{
\item{data}{a dataframe or tibble}

\item{ncol}{specific column, factor}
}
\value{
a figure as ggplot2 object
}
\description{
Using a specific column of a data frame to plot a pie plot-v2
}
\examples{
my_pie2(mtcars,ncol=2)
my_pie2(diamonds,ncol=3)
my_pie2(diamonds,ncol=3)+guides(fill="none")
}
